requirements.txt
absl-py==2.1.0
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
einops==0.8.1
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2025.3.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grad-cam==1.5.4
grpcio==1.70.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
joblib==1.4.2
kagglehub==0.2.9
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft
mkl-random
mkl-service==2.4.0
numpy
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.0.3
pillow
pluggy==1.5.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.15
tomli==2.2.1
tqdm==4.67.1
ttach==0.0.3
typing_extensions
tzdata==2024.2
-e git+https://github.com/ultralytics/ultralytics.git@00aefd795c13bb6a5364b63798c383adc15248d4#egg=ultralytics
ultralytics-thop==2.0.13
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
